src/, lib/, tests/: Rename files defining strtcpy()
