+++ lib/Makefile.am
@@ -141,7 +141,7 @@ libshadow_la_SOURCES = \
 	stpecpy.h \
 	stpeprintf.c \
 	stpeprintf.h \
-	strlcpy.h \
+	strtcpy.h \
 	strtoday.c \
 	sub.c \
 	subordinateio.h \
