+++ lib/strtcpy.h
