+++ src/chage.c
@@ -33,7 +33,7 @@
 #include "pwio.h"
 #include "shadowio.h"
 #include "shadowlog.h"
-#include "strlcpy.h"
+#include "strtcpy.h"
 #ifdef WITH_TCB
 #include "tcbfuncs.h"
 #endif
