+++ src/login.c
@@ -36,7 +36,7 @@
 /*@-exitarg@*/
 #include "exitcodes.h"
 #include "shadowlog.h"
-#include "strlcpy.h"
+#include "strtcpy.h"
 
 #ifdef USE_PAM
 #include "pam_defs.h"
