+++ src/passwd.c
@@ -32,7 +32,7 @@
 #include "pwio.h"
 #include "shadowio.h"
 #include "shadowlog.h"
-#include "strlcpy.h"
+#include "strtcpy.h"
 
 /*
  * exit status values
